Massively improve game UI/UX
The game will now reset when clicking the button. Also, fixed an issue where the game area would duplicate after each click instead of resetting. This was fixed by clearing the grid-container node each time the button is clicked. The game will also start upon page load with a default value of 50 instead of waiting for the user to click the button.